requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
jinja2>=3.1.4,<4.0.0
python-dotenv>=1.0.1,<2.0.0
httpx>=0.28.1,<1.0.0
playwright>=1.55.0,<2.0.0
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.10.0,<3.0.0
pypdf>=5.4.0
gspread>=6.1.0,<7.0.0
google-auth>=2.35.0,<3.0.0
browser-use>=0.12.0
langchain-google-genai>=2.0.0
